- fix a bunch of msvc warnings (mostly assigning double constants to float variables)
SVN-Revision: 4661

+++ src/badguy/flame.cpp
@@ -51,7 +51,7 @@ Flame::write(lisp::Writer& writer)
 void
 Flame::active_update(float elapsed_time)
 {
-  angle = fmodf(angle + elapsed_time * speed, 2*M_PI);
+  angle = fmodf(angle + elapsed_time * speed, (float) (2*M_PI));
   Vector newpos(start_position.x + cos(angle) * radius,
                 start_position.y + sin(angle) * radius);
   movement = newpos - get_pos();
